requirements.txt
pydub==0.25.1
SpeechRecognition==3.8.1
protobuf==3.18.1
google-cloud-speech==2.9.3
kafka-python==2.0.2
pymongo==3.12.0
mako==1.1.5
pytube==11.0.1
